+++ src/main/java/com/refinedmods/refinedstorage/apiimpl/network/Network.java
@@ -525,7 +525,12 @@ public class Network implements INetwork, IRedstoneConfigurable {
     }
 
     private void updateEnergyUsage() {
-        int usage = redstoneMode.isEnabled(world, pos) ? RS.SERVER_CONFIG.getController().getBaseUsage() : 0;
+        if (!redstoneMode.isEnabled(world, pos)) {
+            this.lastEnergyUsage = 0;
+            return;
+        }
+
+        int usage = RS.SERVER_CONFIG.getController().getBaseUsage();
 
         for (INetworkNode node : nodeGraph.all()) {
             if (node.isActive()) {